fix: errors, makefile changed, all test added main

+++ libasm.h
@@ -6,7 +6,7 @@
 /*   By: javrodri <javrodri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/03 19:47:07 by javrodri          #+#    #+#             */
-/*   Updated: 2020/10/01 20:14:57 by javrodri         ###   ########.fr       */
+/*   Updated: 2020/10/05 18:28:21 by javrodri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 #include <errno.h>
 #include <fcntl.h>
 
-int		ft_strlen(char const *str);
+size_t		ft_strlen(char const *str);
 int		ft_strcmp(char const *s1, char const *s2);
-char		*ft_strcpy(char *dst, char const *src);
-ssize_t		ft_write(int fd, void const *buf, size_t nbyte);
-ssize_t		ft_read(int fd, void *buf, size_t nbyte);
-char		*ft_strdup(char const *s1);
+char	*ft_strcpy(char *dst, char const *src);
+size_t		ft_write(int fd, void const *buf, size_t nbyte);
+size_t		ft_read(int fd, void *buf, size_t nbyte);
+char    *ft_strdup(char const *s1);
